refactor: utils 디렉토리 생성
